move `FbVarScreeninfo` and `FbFixScreeninfo` to hal-unix
